combined_service/pyproject.toml
[tool.poetry]
name = "combined_mtg"
version = "0.1.0"
description = ""
authors = ["Ramón Molina", "Javier Martínez"]

[tool.poetry.dependencies]
python = ">3.10.0"
pymongo = "^4.3.3"
fastapi = "^0.88.0"
uvicorn = "0.20.0"
load-dotenv = "^0.1.0"
grpcio = "^1.51.0"

[tool.poetry.dev-dependencies]
black = "^22.10.0"
pytest = "^7.2.0"
tomli = "^2.0.1"
mypy = "^0.991"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
